requirements.txt
flask
google-cloud-storage
google-cloud-secret-manager
google-cloud-logging
google-cloud-trace
google-cloud-vision
python-dotenv
tenacity
cryptography
flask-cors
